refactor(PhraseEditor): break 556-line component into hook + parts
- usePhraseEditor hook owns phrase/metadata loading, inline edit/save/delete,
  add-phrase, and duplicate/metadata update handling
- PhraseEditorHeader, PhraseStatsBar, AddPhraseForm, PhraseListItem extracted
- Add/edit inputs reuse the shared FormField primitive
- Folder with index.tsx orchestrator; 2 import sites unchanged

Verified: tsc clean, lint 0 errors, build succeeds. No behavior change.
